+++ module-infra/src/main/java/com/wsws/moduleinfra/repo/feed/QuestionJpaRepository.java
@@ -1,7 +1,7 @@
 package com.wsws.moduleinfra.repo.feed;
 
 import com.wsws.moduledomain.feed.question.repo.QuestionRepository;
-import com.wsws.moduleinfra.entity.QuestionEntity;
+import com.wsws.moduleinfra.entity.feed.QuestionEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
